+++ src/interfaces/cellTemplateContext/cellTemplate.context.ts
@@ -2,7 +2,7 @@
  * Context for cell template
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export interface CellTemplateContext<TColumnMetadata = any>
+export interface CellTemplateContext<TColumnMetadata = unknown>
 {
     //######################### readonly properties #########################