+++ src/plugins/contentRenderer/bodyHeader/bodyContentRendererAbstract.component.ts
@@ -9,7 +9,7 @@ import {GridPluginInstances} from '../../../misc/types';
  * Abstract component for body content renderer
  */
 @Directive()
-export class BodyContentRendererAbstractComponent<TData = unknown, TCssClasses = unknown, TOptions extends BodyContentRendererOptions<TCssClasses> = BodyContentRendererOptions<TCssClasses>, TMetadata = unknown> implements BodyContentRenderer<TData, TMetadata>, GridPlugin<TOptions>
+export class BodyContentRendererAbstractComponent<TData = unknown, TCssClasses = unknown, TOptions extends BodyContentRendererOptions<TData, TCssClasses> = BodyContentRendererOptions<TData, TCssClasses>, TMetadata = unknown> implements BodyContentRenderer<TData, TMetadata>, GridPlugin<TOptions>
 {
     //######################### protected fields #########################